Fix spec (Follow operationContext removing)
